fixing the alert dialog in ipad and iphone
